installs and configure BrowserRouter
