+++ src/components/Footer.jsx
@@ -1,0 +1,9 @@
+function Footer() {
+  return (
+    <footer className="mt-auto bg-violet-500 py-8 text-center text-white">
+      <p>&copy; eStore. All rights reserved.</p>
+    </footer>
+  );
+}
+
+export default Footer;
