+++ src/components/Layout.jsx
@@ -1,0 +1,15 @@
+import Header from "./Header";
+import Footer from "./Footer";
+import { Outlet } from "react-router-dom";
+
+function Layout() {
+  return (
+    <div>
+      <Header />
+      <Outlet />
+      <Footer />
+    </div>
+  );
+}
+
+export default Layout;
