+++ src/routes/Home.jsx
@@ -1,0 +1,9 @@
+function Home() {
+  return (
+    <div className="">
+      <h1 className="font-bold">This is Home</h1>
+    </div>
+  );
+}
+
+export default Home;
